// frontend/src/api/generated/model/userBasicDTO.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * Your API
 * OpenAPI spec version: 1.0.0
 */

export interface UserBasicDTO {
  id?: number;
  name?: string;
  email?: string;
  phoneNumber?: string;
  enabled?: boolean;
  accountNonLocked?: boolean;
  lockedUntil?: string;
  roleNames?: string[];
  lastLoginTime?: string;
  createdAt?: string;
  updatedAt?: string;
}